Update CHARM.smk, config.yaml, and rules
This commit updates the CHARM.smk file to limit the number of samples to 10. It also modifies the config.yaml file to adjust the number of CPU threads for various processes. Additionally, the rules for splitting, cutting DNA, and cutting RNA have been modified. Finally, the runCHARM.sh script has been updated to adjust the number of jobs and resources.